Extract base test
